Add packages to replace includes.
A package is a folder that contains modules.
A package may also have dependencies, which are other packages.
As opposed to includes, packages may only import files from their own
folder, or their dependencies' folders.

This is necessary to allow disambiguating rebuilding the compiler when
macros are in play. More than that, it's just cleaner.