It is fundamentally wrong to pull in the compiler version of macros!
Consider a quasiquoting macro. The class constructor call that it
*generates*, has to be the one with the semantics of the building compiler.
However, the class that it *deconstructs* has to be the one of the
running compiler. So if you use the compiler's quasiquoting macro,
you'll get a generated constructor call that matches your active
compiler, which is *not* your next compiler, and so the generated
calls will be wrong.