switch to staticAlloca instead of stacksave/alloca/stackrestore pattern.
LLVM is really not happy with all the stacksave/alloca/stackrestore
that I've been doing. It's lead to at least one filed optimizer bug,
and I suspect was about to lead to another one and more after. Plus,
the generated ASM was *really weird* and filled with odd patterns like

  mov rsp, rcx
  mov rcx, rsp

So in exchange for losing the "benefit" of getting different pointers
on every alloca call, a benefit that we then spent extra effort
to get rid of anyway, we sidestep all these issues entirely by moving
all allocas to the function entry. Hence 'staticAlloca'.